+++ apps/frontend/src/_components/pages/main/SignupForm.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import styles from "./FormSection.module.css";
+
+export default function SignupForm() {
+  return <form className={styles.form}></form>;
+}
